chore: Split up `oneil_output::number` module
